mysite/requirements-dev.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.2
idna==2.6
nltk==3.2.5
numpy==1.14.0
oauthlib==2.0.6
python-decouple==3.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
static3==0.7.0
textblob==0.15.1
tweepy==3.5.0
urllib3==1.22
